Another attempt to fix the Windows linking bug
