--- core/base/meshGraph/MeshGraph.cpp
@@ -1,1 +1,0 @@
-// This file needs to exist for the Mac compiler
