+++ core/vtk/ttkAlgorithm/ttk.module
@@ -2,7 +2,6 @@ NAME
   ttkAlgorithm
 SOURCES
   ttkAlgorithm.cpp
-  ttkUtils.cpp
 HEADERS
   ttkAlgorithm.h
   ttkUtils.h
